Reorganizes utils
