+++ atoma-api/src/common/graphql/errors/base.error.ts
