+++ atoma-api/src/common/neo4j-utils/base.repository.ts
@@ -1,7 +1,7 @@
 import { Neo4jService } from '@modules/database/neo.service';
 import { ClassConstructor, plainToInstance } from 'class-transformer';
 import { Query } from './types';
-import { Paginated } from '@common/pagination/pagination.types';
+import { Paginated } from '@common/graphql/pagination/pagination.types';
 
 export abstract class BaseRepository<T> {
   constructor(
