+++ atoma-api/src/common/neo4j-utils/index.ts
@@ -1,3 +1,2 @@
 export * from './base.repository';
-export * from './base.schema';
 export * from './types';
