+++ atoma-api/src/common/neo4j-utils/types.ts
