+++ atoma-api/src/modules/compounds/results/find-compound.result.ts
@@ -1,4 +1,4 @@
-import { NotFoundError } from '@common/errors/not-found.error';
+import { NotFoundError } from '@common/graphql/errors/not-found.error';
 import { createUnionType } from '@nestjs/graphql';
 import { Compound } from '@schemas/compound.schema';
