+++ atoma-api/src/modules/compounds/services/compound-properties.service.ts
@@ -3,8 +3,8 @@ import { CompoundPropertiesRepository } from '../repositories/compound-propertie
 import { CompoundProperty } from '@schemas/compound-property.schema';
 import { Neo4jService } from '@modules/database/neo.service';
 import { v4 as uuidv4 } from 'uuid';
-import { FindPaginatedInput } from '@common/pagination/pagination.input';
-import { Paginated } from '@common/pagination/pagination.types';
+import { FindPaginatedInput } from '@common/graphql/pagination/pagination.input';
+import { Paginated } from '@common/graphql/pagination/pagination.types';
 
 // interface FindOneParms {
 //   compoundUuid: string;
