+++ atoma-api/src/modules/properties/properties.service.ts
@@ -5,11 +5,11 @@ import {
 } from '@nestjs/common';
 import { Property } from '@schemas/property.schema';
 import { PropertiesRepository } from './properties.repository';
-import { NotFoundError } from '@common/errors/not-found.error';
-import { Query } from '@common/repositories';
+import { NotFoundError } from '@common/graphql/errors/not-found.error';
+import { Query } from '@common/neo4j-utils';
 import { v4 as uuidv4 } from 'uuid';
-import { FindPaginatedInput } from '@common/pagination/pagination.input';
-import { Paginated } from '@common/pagination/pagination.types';
+import { FindPaginatedInput } from '@common/graphql/pagination/pagination.input';
+import { Paginated } from '@common/graphql/pagination/pagination.types';
 
 @Injectable()
 export class PropertiesService {
